Revert "Move AbstractLexicoder to API (#1400)"
This reverts commit 2e8144ec89e74d37c7ff97d301051fea747e05c7.
