+++ core/src/main/java/org/apache/accumulo/core/client/lexicoder/BytesLexicoder.java
@@ -18,6 +18,8 @@
  */
 package org.apache.accumulo.core.client.lexicoder;
 
+import org.apache.accumulo.core.clientImpl.lexicoder.AbstractLexicoder;
+
 /**
  * For each of the methods, this lexicoder just passes the input through untouched. It is meant to
  * be combined with other lexicoders like the {@link ReverseLexicoder}.
