+++ core/src/main/java/org/apache/accumulo/core/client/lexicoder/DateLexicoder.java
@@ -20,6 +20,8 @@ package org.apache.accumulo.core.client.lexicoder;
 
 import java.util.Date;
 
+import org.apache.accumulo.core.clientImpl.lexicoder.AbstractLexicoder;
+
 /**
  * A lexicoder for date objects. It preserves the native Java sort order for Date.
  *
