+++ core/src/main/java/org/apache/accumulo/core/client/lexicoder/DoubleLexicoder.java
@@ -18,6 +18,8 @@
  */
 package org.apache.accumulo.core.client.lexicoder;
 
+import org.apache.accumulo.core.clientImpl.lexicoder.AbstractLexicoder;
+
 /**
  * A lexicoder for preserving the native Java sort order of Double values.
  *
