+++ core/src/main/java/org/apache/accumulo/core/client/lexicoder/StringLexicoder.java
@@ -20,6 +20,8 @@ package org.apache.accumulo.core.client.lexicoder;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
+import org.apache.accumulo.core.clientImpl.lexicoder.AbstractLexicoder;
+
 /**
  * This lexicoder encodes/decodes a given String to/from bytes without further processing. It can be
  * combined with other encoders like the {@link ReverseLexicoder} to flip the default sort order.
