+++ core/src/main/java/org/apache/accumulo/core/client/lexicoder/TextLexicoder.java
@@ -18,6 +18,7 @@
  */
 package org.apache.accumulo.core.client.lexicoder;
 
+import org.apache.accumulo.core.clientImpl.lexicoder.AbstractLexicoder;
 import org.apache.accumulo.core.util.TextUtil;
 import org.apache.hadoop.io.Text;
