+++ core/src/main/java/org/apache/accumulo/core/client/lexicoder/ULongLexicoder.java
@@ -18,6 +18,8 @@
  */
 package org.apache.accumulo.core.client.lexicoder;
 
+import org.apache.accumulo.core.clientImpl.lexicoder.AbstractLexicoder;
+
 /**
  * Unsigned long lexicoder. The lexicographic encoding sorts first 0l and -1l last. This encoding
  * does not correspond to the sort of Long because it does not consider the sign bit. If Java had an
