+++ core/src/main/java/org/apache/accumulo/core/clientImpl/lexicoder/AbstractLexicoder.java
@@ -16,6 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.core.client.lexicoder;
+package org.apache.accumulo.core.clientImpl.lexicoder;
+
+import org.apache.accumulo.core.client.lexicoder.AbstractEncoder;
+import org.apache.accumulo.core.client.lexicoder.Lexicoder;
 
 public abstract class AbstractLexicoder<T> extends AbstractEncoder<T> implements Lexicoder<T> {}
